cleaning up spack and conda environment classes
there is a lot of shared functionality that now is represented in the base class.
I have also added a EnvCommandBase that can have shared functionality to look for the environment command (e.g., spack or conda)
